# Improve the performance with over 20% for Go by using uint