Hopefully, using uint and uint32 are within the rules (I could not find any particular rules for the program in README.md).

Here is the new benchmark result:

```
❯ hyperfine -w 1 -r 2 ./main
Benchmark 1: ./main
  Time (mean ± σ):      1.665 s ±  0.023 s    [User: 1.655 s, System: 0.004 s]
  Range (min … max):    1.649 s …  1.681 s    2 runs
```

Compared to previously:

```
❯ hyperfine -w 1 -r 2 ./main
Benchmark 1: ./main
  Time (mean ± σ):      2.150 s ±  0.014 s    [User: 2.139 s, System: 0.005 s]
  Range (min … max):    2.140 s …  2.160 s    2 runs
```

Also, `result[0] = 0` is not needed, since arrays are initialized to have 0 values in Go.